+++ src/components/AboutMe/AboutMe.jsx
@@ -1,0 +1,9 @@
+import React, { useState } from "react";
+import "./AboutMe.css";
+export default function NavBar() {
+  return (
+    <>
+      <div className="about-me"></div>
+    </>
+  );
+}
