+++ src/components/Project/project.css
@@ -1,5 +1,4 @@
 .mainProject{
-    background-color: rgba(0, 0, 0, 0.25);
     border-radius: 20px;
     margin: 3%;
     padding: 3%;
