+++ src/components/Socials/socials.css
@@ -1,6 +1,5 @@
 .social{
     display: flex;
-    background-color: rgba(0, 0, 0, 0.185);
     border-radius: 20px; 
     text-align: center;
     margin: 5%;
