+++ src/components/TechStack/techStack.css
@@ -1,6 +1,5 @@
 .techstack{
     display: flex;
-    background-color: rgba(0, 0, 0, 0.185);
     border-radius: 20px;
     margin: 5%;
     padding: 5%;
