+++ src/index.css
@@ -7,12 +7,6 @@
   font-family: "Fjalla";
   src: url("./fonts/FjallaOne-Regular.ttf") format("truetype");
 }
-body {
-  background: 
-  linear-gradient(#e66465, #9198e5);
-  
-  background-size: cover;
-  font-family: 'Fjalla', sans-serif;
-}
+
