+++ src/pages/HomePage/HomePage.css
@@ -24,20 +24,17 @@
 }
 .aboutMe{
     margin: 5%;
-    background-color: rgba(0, 0, 0, 0.185);
     border-radius: 20px;
     padding: 2%;
 }
 .projects{
     margin: 5%;
-    background-color: rgba(0, 0, 0, 0.185);
     border-radius: 20px;
     padding: 1%;
 
 }
 .experience{
     margin: 5%;
-    background-color: rgba(0, 0, 0, 0.185);
     border-radius: 20px;
     padding: 1%;
 
@@ -72,15 +69,23 @@
     animation: spin-anim 2s linear infinite;
     
 }
-@keyframes spin-anim {
-    0% {
-        transform: rotate(0deg);
-    }
-    100% {
-        transform: rotate(360deg);
-    }
+
+
+
+.home-page__container{
+    z-index: 30;
 }
-.experienceList{
-    text-align: left;
-    margin: 5%;
+.home-page__container--content{
+    display: flex;
+    z-index: 30;
+    
 }
+.home-page__container--content-left{
+    width: 40%;
+    z-index: 20;
+}
+
+.home-page__container--content-right{
+    width: 60%;
+}
+
